lcs-sacoche_trunk : update to version 2014-06-26
git-svn-id: https://svn.tice.ac-caen.fr/svn/LcsSE3/LCS/LCS2.0ndg/sources/trunk@5088 1012e355-7652-49d3-8fd6-1f9b0b464cea

+++ lcs-sacoche/Sacoche/_sql/structure/sacoche_niveau.sql
@@ -1,7 +1,7 @@
 DROP TABLE IF EXISTS sacoche_niveau;
 
 CREATE TABLE sacoche_niveau (
-  niveau_id         TINYINT(3)  UNSIGNED                NOT NULL AUTO_INCREMENT,
+  niveau_id         SMALLINT(5) UNSIGNED                NOT NULL AUTO_INCREMENT,
   niveau_actif      TINYINT(1)  UNSIGNED                NOT NULL DEFAULT 0,
   niveau_famille_id TINYINT(3)  UNSIGNED                NOT NULL DEFAULT 0,
   niveau_ordre      TINYINT(3)  UNSIGNED                NOT NULL DEFAULT 0,
